autodetect gdc/ldc2 D-compiler, and *.d portability cleanups

+++ Data_Generators/makedata/logmake.d
@@ -145,7 +145,7 @@ char [][]trimouterblanks(char [][]input) {
 void processlogs() {
 	char [][]output;
 	//printf ("Processing logs\n");
-	foreach(int i, Log log; loglist) {
+	foreach(size_t i, Log log; loglist) {
 		//printf("%d:%d:[%s]\n", log.id, log.title.length, log.title.toStringz);
 		log.head = wraplines(log.head, 49);
 		//printf(".\n");
